ai-resume-ats/backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
scikit-learn==1.3.0
spacy==3.6.1
PyPDF2==3.0.1
python-docx==0.8.11
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3
gunicorn==21.2.0